Update read.me
Each section is a progression from the experience of the previous.
There is overlap in terms of using previous scripts with more data, for the purpose of exploring accuracy, precision, recall and TP/TF ratios.
Section 5 is the FAIR metadata  and is the XML on the G4HSP-AI directory rather than  a script module.

Please consider this as "exploration" .
There is a new and final update that will include an interactive knowledge graph with SparQL and an interactive AI for searching virus G4-HSP interaction. This will be released as  new journal publication.